+++ src/graph.h
@@ -13,8 +13,8 @@ typedef struct Graph
   // The number of edges and vertices in the graph.
   unsigned int numedges, numvertices;
   int*  vertexIds;
-  double* vertexXLocs;
-  double* vertexYLocs;
+  float* vertexXLocs;
+  float* vertexYLocs;
   unsigned int* edgeSources;
   unsigned int* edgeTargets;
 } Graph;
